Split implementation of topological operators into multiple files (#992)
