+++ libs/vgc/vacomplex/detail/move.cpp
@@ -1,0 +1,143 @@
+// Copyright 2022 The VGC Developers
+// See the COPYRIGHT file at the top-level directory of this distribution
+// and at https://github.com/vgc/vgc/blob/master/COPYRIGHT
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include <vgc/vacomplex/detail/operations.h>
+
+namespace vgc::vacomplex::detail {
+
+void Operations::insertNodeBeforeSibling_(Node* node, Node* nextSibling) {
+    Group* oldParent = node->parentGroup();
+    Group* newParent = nextSibling->parentGroup();
+    if (newParent->insertChildUnchecked(nextSibling, node)) {
+        onNodeInserted_(node, oldParent, NodeInsertionType::BeforeSibling);
+    }
+}
+
+void Operations::insertNodeAfterSibling_(Node* node, Node* previousSibling) {
+    Group* oldParent = node->parentGroup();
+    Group* newParent = previousSibling->parentGroup();
+    Node* nextSibling = previousSibling->nextSibling();
+    if (newParent->insertChildUnchecked(nextSibling, node)) {
+        onNodeInserted_(node, oldParent, NodeInsertionType::AfterSibling);
+    }
+}
+
+void Operations::insertNodeAsFirstChild_(Node* node, Group* parent) {
+    Group* oldParent = node->parentGroup();
+    Node* nextSibling = parent->firstChild();
+    if (parent->insertChildUnchecked(nextSibling, node)) {
+        onNodeInserted_(node, oldParent, NodeInsertionType::FirstChild);
+    }
+}
+
+void Operations::insertNodeAsLastChild_(Node* node, Group* parent) {
+    Group* oldParent = node->parentGroup();
+    if (parent->appendChild(node)) {
+        onNodeInserted_(node, oldParent, NodeInsertionType::LastChild);
+    }
+}
+
+void Operations::moveToGroup(Node* node, Group* parentGroup, Node* nextSibling) {
+    if (nextSibling) {
+        insertNodeBeforeSibling_(node, nextSibling);
+    }
+    else {
+        insertNodeAsLastChild_(node, parentGroup);
+    }
+}
+
+void Operations::moveBelowBoundary(Node* node) {
+    Cell* cell = node->toCell();
+    if (!cell) {
+        return;
+    }
+    auto boundary = cell->boundary();
+    if (boundary.isEmpty()) {
+        // nothing to do.
+        return;
+    }
+    // currently keeping the same parent
+    Node* oldParentNode = cell->parentGroup();
+    Node* newParentNode = oldParentNode;
+    if (!newParentNode) {
+        // `boundary.length() > 0` previously checked.
+        newParentNode = (*boundary.begin())->parentGroup();
+    }
+    if (!newParentNode) {
+        return;
+    }
+    Group* newParent = newParentNode->toGroupUnchecked();
+    Node* nextSibling = newParent->firstChild();
+    while (nextSibling) {
+        bool found = false;
+        for (Cell* boundaryCell : boundary) {
+            if (nextSibling == boundaryCell) {
+                found = true;
+                break;
+            }
+        }
+        if (found) {
+            break;
+        }
+        nextSibling = nextSibling->nextSibling();
+    }
+    if (nextSibling) {
+        insertNodeBeforeSibling_(node, nextSibling);
+    }
+    else {
+        // all boundary cells are in another group
+        // TODO: use set of ancestors of boundary cells
+        insertNodeAsLastChild_(node, newParent);
+    }
+}
+
+Node* Operations::findTopMost(core::ConstSpan<Node*> nodes) {
+    // currently only looking under a single parent
+    // TODO: tree-wide top most.
+    if (nodes.isEmpty()) {
+        return nullptr;
+    }
+    Node* node0 = nodes[0];
+    Group* parent = node0->parentGroup();
+    Node* topMostNode = parent->lastChild();
+    while (topMostNode) {
+        if (nodes.contains(topMostNode)) {
+            break;
+        }
+        topMostNode = topMostNode->previousSibling();
+    }
+    return topMostNode;
+}
+
+Node* Operations::findBottomMost(core::ConstSpan<Node*> nodes) {
+    // currently only looking under a single parent
+    // TODO: tree-wide bottom most.
+    if (nodes.isEmpty()) {
+        return nullptr;
+    }
+    Node* node0 = nodes[0];
+    Group* parent = node0->parentGroup();
+    Node* bottomMostNode = parent->firstChild();
+    while (bottomMostNode) {
+        if (nodes.contains(bottomMostNode)) {
+            break;
+        }
+        bottomMostNode = bottomMostNode->nextSibling();
+    }
+    return bottomMostNode;
+}
+
+} // namespace vgc::vacomplex::detail
