+++ libs/vgc/vacomplex/detail/unglue.cpp
@@ -1,0 +1,304 @@
+// Copyright 2022 The VGC Developers
+// See the COPYRIGHT file at the top-level directory of this distribution
+// and at https://github.com/vgc/vgc/blob/master/COPYRIGHT
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include <vgc/vacomplex/detail/operations.h>
+
+namespace vgc::vacomplex::detail {
+
+Int Operations::countUses_(KeyVertex* kv) {
+    Int count = 0;
+    for (Cell* starCell : kv->star()) {
+        switch (starCell->cellType()) {
+        case CellType::KeyEdge: {
+            KeyEdge* ke = starCell->toKeyEdgeUnchecked();
+            bool hasFaceInStar = false;
+            for (Cell* keStarCell : ke->star()) {
+                if (keStarCell->cellType() == CellType::KeyFace) {
+                    hasFaceInStar = true;
+                    break;
+                }
+            }
+            if (!hasFaceInStar) {
+                if (ke->isStartVertex(kv)) {
+                    ++count;
+                }
+                if (ke->isEndVertex(kv)) {
+                    ++count;
+                }
+            }
+            break;
+        }
+        case CellType::KeyFace: {
+            KeyFace* kf = starCell->toKeyFaceUnchecked();
+            for (const KeyCycle& cycle : kf->cycles()) {
+                if (cycle.steinerVertex()) {
+                    if (cycle.steinerVertex() == kv) {
+                        ++count;
+                    }
+                    continue;
+                }
+                for (const KeyHalfedge& khe : cycle.halfedges()) {
+                    if (khe.startVertex() == kv) {
+                        ++count;
+                    }
+                }
+            }
+            break;
+        }
+        default:
+            break;
+        }
+    }
+    return count;
+}
+
+Int Operations::countUses_(KeyEdge* ke) {
+    Int count = 0;
+    for (Cell* starCell : ke->star()) {
+        KeyFace* kf = starCell->toKeyFace();
+        if (!kf) {
+            continue;
+        }
+        for (const KeyCycle& cycle : kf->cycles()) {
+            if (cycle.steinerVertex()) {
+                continue;
+            }
+            for (const KeyHalfedge& khe : cycle.halfedges()) {
+                if (khe.edge() == ke) {
+                    ++count;
+                }
+            }
+        }
+    }
+    return count;
+}
+
+core::Array<KeyEdge*> Operations::unglueKeyEdges(KeyEdge* targetKe) {
+    core::Array<KeyEdge*> result;
+    if (countUses_(targetKe) <= 1) {
+        result.append(targetKe);
+        return result;
+    }
+
+    // TODO: handle temporal star.
+
+    // Helper
+    auto duplicateTargetKe = [this, targetKe, &result]() {
+        KeyEdge* newKe = nullptr;
+        KeyEdgeData dataDuplicate = targetKe->data();
+        if (targetKe->isClosed()) {
+            newKe = createKeyClosedEdge(
+                std::move(dataDuplicate),
+                targetKe->parentGroup(),
+                targetKe->nextSibling());
+        }
+        else {
+            newKe = createKeyOpenEdge(
+                targetKe->startVertex(),
+                targetKe->endVertex(),
+                std::move(dataDuplicate),
+                targetKe->parentGroup(),
+                targetKe->nextSibling());
+        }
+        result.append(newKe);
+        return newKe;
+    };
+
+    // Substitute targetKe by a duplicate in each of its use.
+    // Note: star is copied for safety since it may be modified in loop.
+    for (Cell* cell : targetKe->star().copy()) {
+        switch (cell->cellType()) {
+        case CellType::KeyFace: {
+            KeyFace* kf = cell->toKeyFaceUnchecked();
+            for (KeyCycle& cycle : kf->cycles_) {
+                if (cycle.steinerVertex_) {
+                    continue;
+                }
+                KeyHalfedge first = cycle.halfedges().first();
+                if (!first.isClosed()) {
+                    for (KeyHalfedge& kheRef : cycle.halfedges_) {
+                        if (kheRef.edge() == targetKe) {
+                            KeyEdge* newKe = duplicateTargetKe();
+                            kheRef = KeyHalfedge(newKe, kheRef.direction());
+                            addToBoundary_(kf, newKe);
+                        }
+                    }
+                    VGC_ASSERT(cycle.isValid());
+                }
+                else if (first.edge() == targetKe) {
+                    KeyEdge* newKe = duplicateTargetKe();
+                    for (KeyHalfedge& kheRef : cycle.halfedges_) {
+                        kheRef = KeyHalfedge(newKe, kheRef.direction());
+                    }
+                    addToBoundary_(kf, newKe);
+                    // TODO: instead of having a copy of the edge used N times
+                    // use a single edge with its geometry looped N times.
+                    // See Boris Dalstein's thesis page 187.
+                    VGC_ASSERT(cycle.isValid());
+                }
+            }
+            removeFromBoundary_(kf, targetKe);
+            break;
+        }
+        default:
+            throw core::LogicError(
+                "unglueKeyEdges() doesn't support temporal cells in edge star.");
+        }
+    }
+
+    // Delete targetKe
+    hardDelete(targetKe);
+
+    return result;
+}
+
+core::Array<KeyVertex*> Operations::unglueKeyVertices(
+    KeyVertex* targetKv,
+    core::Array<std::pair<core::Id, core::Array<KeyEdge*>>>& ungluedKeyEdges) {
+
+    core::Array<KeyVertex*> result;
+    if (countUses_(targetKv) <= 1) {
+        result.append(targetKv);
+        return result;
+    }
+
+    // TODO: handle temporal star.
+
+    // Unglue incident key edges.
+    for (Cell* cell : targetKv->star().copy()) {
+        switch (cell->cellType()) {
+        case CellType::KeyEdge: {
+            KeyEdge* ke = cell->toKeyEdgeUnchecked();
+            core::Id id = ke->id();
+            core::Array<KeyEdge*> a = unglueKeyEdges(ke);
+            if (a.length() > 1) {
+                ungluedKeyEdges.emplaceLast(id, std::move(a));
+            }
+            break;
+        }
+        default:
+            break;
+        }
+    }
+
+    // Helper
+    auto duplicateTargetKv = [this, targetKv, &result]() {
+        KeyVertex* newKv = createKeyVertex(
+            targetKv->position(),
+            targetKv->parentGroup(),
+            targetKv->nextSibling(),
+            targetKv->time());
+        result.append(newKv);
+        return newKv;
+    };
+
+    // Helper. Assumes the replaced key vertex is `targetKv`.
+    auto substituteTargetKvAtStartOrEndOfKhe = //
+        [this, targetKv](const KeyHalfedge& khe, bool startVertex, KeyVertex* newKv) {
+            //
+            KeyEdge* ke = khe.edge();
+
+            KeyVertex* otherEndKv = nullptr;
+            if (khe.direction() == startVertex) {
+                otherEndKv = ke->endVertex();
+                ke->startVertex_ = newKv;
+            }
+            else {
+                otherEndKv = ke->startVertex();
+                ke->endVertex_ = newKv;
+            }
+
+            if (otherEndKv != targetKv) {
+                removeFromBoundary_(ke, targetKv);
+            }
+
+            addToBoundary_(ke, newKv);
+        };
+
+    // Substitute targetKv by a duplicate in each of its use.
+    // Note: star is copied for safety since it may be modified in loop.
+    for (Cell* cell : targetKv->star().copy()) {
+        switch (cell->cellType()) {
+        case CellType::KeyEdge: {
+            KeyEdge* ke = cell->toKeyEdgeUnchecked();
+            bool hasFaceInStar = false;
+            for (Cell* keStarCell : ke->star()) {
+                if (keStarCell->cellType() == CellType::KeyFace) {
+                    hasFaceInStar = true;
+                    break;
+                }
+            }
+            if (!hasFaceInStar) {
+                if (ke->isStartVertex(targetKv)) {
+                    KeyVertex* newKv = duplicateTargetKv();
+                    ke->startVertex_ = newKv;
+                    addToBoundary_(ke, newKv);
+                }
+                if (ke->isEndVertex(targetKv)) {
+                    KeyVertex* newKv = duplicateTargetKv();
+                    ke->endVertex_ = newKv;
+                    addToBoundary_(ke, newKv);
+                }
+                removeFromBoundary_(ke, targetKv);
+            }
+            break;
+        }
+        case CellType::KeyFace: {
+            KeyFace* kf = cell->toKeyFaceUnchecked();
+            for (KeyCycle& cycle : kf->cycles_) {
+                if (cycle.steinerVertex()) {
+                    if (cycle.steinerVertex() == targetKv) {
+                        KeyVertex* newKv = duplicateTargetKv();
+                        cycle.steinerVertex_ = newKv;
+                        addToBoundary_(kf, newKv);
+                    }
+                    continue;
+                }
+                Int numHalfedges = cycle.halfedges_.length();
+                // substitute at face corner uses
+                for (Int i = 0; i < numHalfedges; ++i) {
+                    KeyHalfedge& khe1 = cycle.halfedges_[i];
+                    if (khe1.startVertex() == targetKv) {
+                        Int previousKheIndex = (i - 1 + numHalfedges) % numHalfedges;
+                        KeyHalfedge& khe0 = cycle.halfedges_[previousKheIndex];
+
+                        // (?)---khe0-->(targetKv)---khe1-->(?)
+                        KeyVertex* newKv = duplicateTargetKv();
+                        substituteTargetKvAtStartOrEndOfKhe(khe0, false, newKv);
+                        substituteTargetKvAtStartOrEndOfKhe(khe1, true, newKv);
+                        // (?)---khe0-->( newKv  )---khe1-->(?)
+
+                        addToBoundary_(kf, newKv);
+                    }
+                }
+                VGC_ASSERT(cycle.isValid());
+            }
+            removeFromBoundary_(kf, targetKv);
+            break;
+        }
+        default:
+            throw core::LogicError(
+                "unglueKeyVertices() doesn't support temporal cells in edge star.");
+        }
+    }
+
+    // Delete targetKv
+    hardDelete(targetKv);
+
+    return result;
+}
+
+} // namespace vgc::vacomplex::detail
