+++ libs/vgc/vacomplex/operations.cpp
@@ -16,7 +16,7 @@
 
 #include <vgc/vacomplex/operations.h>
 
-#include <vgc/vacomplex/detail/operationsimpl.h>
+#include <vgc/vacomplex/detail/operations.h>
 
 namespace vgc::vacomplex {
